Remove obsolete implementation
