+++ package.json
@@ -8,10 +8,10 @@
     "dev:electron": "pnpm --filter @agentkib/desktop dev:electron",
     "build": "pnpm --filter @agentkib/desktop build",
     "dist:electron": "pnpm --filter @agentkib/desktop dist:electron",
-    "format": "oxfmt apps/desktop/src",
-    "format:check": "oxfmt --check apps/desktop/src",
-    "lint": "oxlint apps/desktop/src",
-    "lint:fix": "oxlint apps/desktop/src --fix",
+    "format": "oxfmt apps/desktop/src apps/desktop/electron apps/desktop/vite*.config.ts",
+    "format:check": "oxfmt --check apps/desktop/src apps/desktop/electron apps/desktop/vite*.config.ts",
+    "lint": "oxlint apps/desktop/src apps/desktop/electron",
+    "lint:fix": "oxlint apps/desktop/src apps/desktop/electron --fix",
     "test": "pnpm --filter @agentkib/desktop test",
     "typecheck": "pnpm --filter @agentkib/desktop typecheck"
   },
